+++ lib/digests/sha1.dart
@@ -65,7 +65,7 @@ class SHA1Digest extends MD4FamilyDigest implements Digest {
     }
   }
 
-  void processLength(int bitLength) {
+  void processLength(Uint64 bitLength) {
     if( _xOff>14 ) {
       processBlock();
     }
