+++ lib/digests/sha256.dart
@@ -72,7 +72,7 @@ class SHA256Digest extends MD4FamilyDigest implements Digest {
     }
   }
 
-  void processLength(int bitLength) {
+  void processLength(Uint64 bitLength) {
     if( _xOff>14 ) {
       processBlock();
     }
